fix(agent): Removes deprecation warnings
For PHP >= 8.1 deprecation warnings were being generated by the agent
when parsing arguments to the mysqli::real_connect() function wrapper.
This is a combination of the argument type specification string
enabling checking for null values and PHP 8.1 and later outputing a
deprecation warning if a null is provided in this scenario.
The fix is to change the format specification string to the same as what
is used by mysqli::real_connect() for PHP >= 8.1.