feat: add functionality to group errors (#719)
Implement Error Fingerprinting functionality in the PHP Agent.

Error Fingerprinting allows a user to call an Agent API
`newrelic_set_error_group_callback` and provide it a PHP callback
function as an argument.

If the API call is successful, the Agent will register this PHP callback
on a per-request basis and invoke it if the user application encounters
a PHP error. The contents of the callback are completely at the user's
discretion, but it must adhere to the following paradigm:
- The callback must accept *two* arguments that will be PHP arrays
provided by the Agent: The first contains transaction data, the second
contains error data.
- The callback must return a string value that is the error fingerprint
name the user wishes to group the error under in their Errors Inbox.

If successful, an agent attribute `error.group.name` will be sent to the
backend with the return value from the callback.