feat: optional instrumentated function metric (#730)
There still exists a default wrapper, `nr_php_wrap_user_function`, which
(for OAPI) creates an 'after' wraprec that is non-transient and creates
an `InstrumentedFunction` metric. Before this PR, there was the option
to create 'before', 'after', and 'clean' wrappers and define the
transience.

This PR extends the options available when creating a wraprec to no
longer create the InstrumentedFunction metric. It wraps this option with
the other bool option (transience) into an options struct.

Additionally, if the top-most default, `nr_php_wrap_user_function`, is
not used, every single option must be defined ('before', 'after',
'clean', transience, and instrumented metric). This takes the concept
that if a single one of these fields is non-default, it is much more
likely that other fields will also be non-default. The code writer will
be forced to explicitly consider each option available.

This does not add the option to create the InstrumentedFunction metric
when calling `nr_wrap_callable`.