fix(agent): Improves decoration function handling
The decorating processor function was being created multiple times if
more than one handler was pushed.  This was due to the scoped name
of the function (in lowercase) was not being used in the search to see
if it had been previously created.  A new integration test was added
to check the use case where more than one handler is being used.

The handler is now checked to see if it supported the pushProcessor()
functionality and if it does not then a warning is generated to notify
the user messages sent to that handler will not be decorated.