fix user exception handler implementation
User exception handler, set with `set_exception_handler` is a callable that
needs to accept one parameter, which will be the Throwable object that was
thrown. If that callable is not defined according to these rules, PHP's
behavior, and therefore agent's behavior, is undefined. E.g. in PHPs 8.0, 8.1
and 8.2, when opcache is disabled Zend Engine gives the agent a zval which is
an object of Throwable type. The same thing happens in PHPs 8.0 and 8.1 when
opcache is enabled. However, when opcache is enabled in PHP 8.2, Zend Engine
no longer gives the agent a zval which is an object of Throwable type.