fix user error handler implementation
User error handler, set with `set_error_handler` is a callable that needs to
accept at least two parameters: int errno and string errstr set to error level
and error message respectively. If that callable is not defined according to
these rules, PHP's behavior, and therefore agent's behavior, is undefined.