tests: fix user exception and error handler implementation (#748)
* fix user exception handler implementation
User exception handler, set with set_exception_handler is a callable that needs
to accept one parameter, which will be the Throwable object that was thrown. If
that callable is not defined according to these rules, PHP's behavior, and
therefore agent's behavior, is undefined. E.g. in PHPs 8.0, 8.1 and 8.2, when
opcache is disabled Zend Engine gives the agent a zval which is an object of
Throwable type. The same thing happens in PHPs 8.0 and 8.1 when opcache is
enabled. However, when opcache is enabled in PHP 8.2, Zend Engine no longer
gives the agent a zval which is an object of Throwable type.

* fix user error handler implementation
User error handler, set with `set_error_handler` is a callable that needs to
accept at least two parameters: int errno and string errstr set to error level
and error message respectively. If that callable is not defined according to
these rules, PHP's behavior, and therefore agent's behavior, is undefined.

* split integration test
An agent hooking into Zend Engine using zend_execute (pre-oapi) hook behaves
differently than the agent hooking into Zend Engine using observer API (oapi).
Namely pre-oapi agent does not generate error events when an exception is
handled with user exception handler wheras oapi agent generates error events in
such case. Therefore the test testing agent's behavior when an exception is
handled by user exception handler must be split depending on what method is
used to hook into Zend Engine: zend_execute hook is used for PHPs < 8.0 and
observer API is used for PHPs >= 8.0.