fix(agent): Changes how log attributes are named
The agent spec specifies that log context attributes name SHOULD be
prefiexed with "context." to avoid name collisions.

This commit changes the agent behavior so log attributes keep the
original name until JSON is being created.  This change allows the
existing attribute filtering mechanism to work with include/exclude
rules which do NOT have to include the "context." prefix.