fix(agent): Fixes log attributes include rule
For log context attributes the filter "include" rule acts exclusively,
differently than existing attributes.  This patch include code to
correct the filter behavior for log context attributes.  This is
accomplished by adding a "finalize rule" as necessary for
the log attribute destination to cause attributes not included
by "include" rules to be dropped.  The exception is if no
include rules are present, or an include rule of "*" is present.