refactor: wrap wp hook callbacks in add_filter
In PHPs 7.4+ use add_filter wrapper to wrap wordpress hooks callbacks.
Older PHPs still use call_user_function_array callback. This approach
limits the number of times the agent is called by Zend Engine during
WordPress transaction execution because add_filter (and add_action,
which is just a wrapper around add_filter) is called less frequently
(called when filter is added to a hook) than call_user_function_array
(called each time hook executes). The number of times the agent is
called by Zend Engine translates to the number of times a wraprec for
user function is looked up, and since the agent changed the method how
it stores wraprecs in PHPs 7.4+ it is an important optimization that
results in lower agent's overhead.