add EXPECT_METRICS_EXIST to integration tests
Integration tests sometimes don't care about the exact set of metrics that
are being sent (particularly around things like datastore metrics); what
is really interesting is that the framework is detected, the transaction
is named correctly, and that there aren't errors. This adds a new
`EXPECT_METRICS_EXIST` section which is simply a newline separated list
of metric names that will be searched for: if any don't exist, the test
will fail.

For example:

```php
<?php

/*EXPECT_METRICS_EXIST
Supportability/framework/WordPress/detected
*/
?>
```

This asserts that the test will generate
`Supportability/framework/WordPress/detected` metric, but doesn't assert
anything else about the metric table.

---------

Co-authored-by: Adam Harvey <adam@adamharvey.name>