use `EXPECT_SPAN_EVENTS_LIKE` in guzzle tests
guzzle integration tests are flaky - they expect span for call to`new Client();`
however when `transaction_tracer.detail` is set to 0, agent sometimes drops it.
On the other hand when `transaction_tracer.detail` is set to 1, the agent
generates more spans. Take advantage of `EXPECT_SPAN_EVENTS_LIKE` and only look
for limited number of spans: root and external.