tests: fix guzzle integration tests (#772)
Guzzle integration tests are flaky - they expect a span for a call to 
`new GuzzleHttp\Client();`.  However, when `transaction_tracer.detail`
is set to 0, the agent sometimes drops it.  On the other hand when
`transaction_tracer.detail` is set to 1, the agent generates many more 
spans and adding them to `EXPECT_SPAN_EVENTS` seems tedious. Therefore
`EXPECT_SPAN_EVENTS_LIKE` is used in test expectations to check if the 
required spans - root and external - are present.