force non-zero duration of helper user function
When tests execute fast enough, the helper function's segment start time and
end time are equal and this causes the segment to be dropped. However, some
integration tests have inconsistent test expectation in expected transaction
traces. On one hand, wildcard match is used for trace details but on the other
hand segment names are expected to match exactly. On faster hardware this
discrepancy will cause such tests to fail. To address this test instability,
non-zero duration is forced on test helper user functions.