improve performance of wrap_generic_callable
nr_php_wrap_generic_callable is used to create transient wraprecs for callables
which don't require to be named. This assumption makes it possible to stay DRY
and reuse nr_php_zval_to_function in lieu of complex pattern matching for the
type of callable. On top of that it improves performance as it now uses
nr_php_wrap_callable always instead of sometimes using nr_php_wrap_user_function