add EXPECT_METRICS_DONT_EXIST to integration tests
Integration tests sometimes don't care about the exact set of metrics that
are being sent (particularly around things like datastore metrics); what
is really interesting is that the framework is detected, the transaction
is named correctly, and that there aren't errors. It is also sometimes
interesting not to see some metrics when a particular feature is disabled.
This adds a new `EXPECT_METRICS_DONT_EXIST` section which is simply a newline
separated list of metric names that will be searched for: if any exist, the
test will fail.

For example:

```php
<?php

/*EXPECT_METRICS_DONT_EXIST
Framework/WordPress/Hook/template_include
*/
?>
```

This asserts that the test will not generate
`Framework/WordPress/Hook/template_include` metric, but doesn't assert
anything else about the metric table.