feat: add go.work to root directory
golang expects the project to be open in the same directory as the
module. This is not the case for the php agent, which operates with
agent/axiom/daemon as sibling directories. In order to make go
understand where the module is, add a go.work file specifying the module
directory path.