feat(oapi): rework exception handling (#767)
Removing the need to manual handling of dangling segments, because Zend
calls all of the hooks we need.

- adds a boolean argument to nr_php_error_record_exception to control
whether we add the error to the current segment. This is needed because
the OAPI context of when the above is called is no longer during a
segment with an uncaught exception and was incorrectly adding the error
to the root segment.
- removes all language of "dangling segments". These no longer exist.
Zend calls all of the necessary `fcall_end`'s, even when an exception is
thrown. When this happens,` func_return_value` is a C `NULL` pointer
which is distinct from a `NULL` (but valid) zval when there is no return
value from a function. We use this `NULL` value to determine the
presence of an uncaught exception.
- no longer overwrites the exception hook; no longer stores a copy of
exceptions locally
- replace storing metadata in the segment, which was used to pair
segments from func_begin with func_end, with logic that always creates
segment in func_begin

Mostly undoes the work of
https://github.com/newrelic/newrelic-php-agent/pull/580

---------

Co-authored-by: Michal Nowacki <mnowacki@newrelic.com>
Co-authored-by: bduranleau-nr <106178551+bduranleau-nr@users.noreply.github.com>
Co-authored-by: Amber Sistla <asistla@newrelic.com>