tests: improve EXPECT_PHP_PACKAGES processing (#827)
This PR includes the following:
- Removes the "options" option which was used to specify that detected
packages would not have version info. This was applied to all packages
detected and was too limiting as some test cases had some packages with
versions and others without (laravel for example). This has been
replaced with the "package_names_only" options which allows giving a
comma separated list of packages which only have names detected.
- Adds checks to verify only one of "supported_packages" or
"expected_packages" is given.
- Improves how "expected_packages" was implemented to overcome some
internal issues that existed.

---------

Co-authored-by: Hitesh Ahuja <108540135+hahuja2@users.noreply.github.com>