fix(testing): Fixes tests for error linking changes
The error linking changes added a "guid" intrinsic attribute to
ErrorTrace and TransactionError JSON.  This patch addresses
these changes.

For tests in the "errors" directory the attributes are tested
explicitely, while in the remaining integration tests which
use EXPECT_TRACED_ERRORS or EXPECT_ERROR_EVENTS
it was left as a wildcard match.  The rationale is the "errors"
tests should provide adequate coverage to ensure the "guid"
attribute is being generated properly.