feat(agent): Always link error traces using transaction ID (#812)
This PR adds the transaction ID to all error traces, even when CAT and
DT are disabled.
This is added in two places:
   - the "CatGUID" field of Error Traces
- as a "guid" attribute in the intrinsic attributes of both Error Traces
and Transaction Errors.
 
It takes advantage of the fact that a random transaction ID is generated
when all transactions are started independent of the DT/CAT settings.

Many integration tests were updated to verify the transaction ID field
and "guid" attributes is now populated.

Additional tests were created to verify the value of the transaction ID
is correct. This required adding some internal testing APIs which give
access to internal agent state needed for the integration test. These
internal testing APIs are not built into release artifacts.