feat(tests): Adds override for package version (#849)
Adds a "override_versions_file" option for EXPECT_PHP_PACKAGE that
allows specifying overrides for given expected versions detected. This
is useful for when a package did not modify the internal data structure
used to by the agent to detect its version, causing the package test to
fail.

The override file should be in the same directory as the
test_php_package.php test and is a JSON file implementing a simple
dictionary like:

{
    "4.13.0": "4.12.0"
}

In this case this specifies if the packge version "4.13.0" is detected
then override this expectation with the version "4.12.0". In this
example the upstream package bumped up to 4.13.0 but did not change the
internal version variable so the agent still detects it as 4.12.0.