fix recording errors for PHPs 8.0+
When user exception handler is installed it can call restore_exception_handler,
and that will reset is_exception_handler flag in the wraprec which will cause
the agent not to use the correct code path after user exception handler returns.
The information that `nr_php_instrument_func_end` is handling return from user
exception handler needs to be stored in the segment associated with the user
exception handler.