fix(agent): Fixes warning from mysqli when explaining slow SQL queries (GH issue 875) (#881)
In GH issue #875 it was reported a warning is generated when using
prepared statements with `mysqli`. The issue included excellent
instructions on how to reproduce the issue. This PR addresses the cause
of the warning.

To understand the cause of the warning requires understanding how the
agent handles `mysqli::stmt` objects. When a SQL statement is prepared
via mysqli the agent will store the SQL string in a global hashmap
called `mysqli_queries` *if* the SQL is considered explainable
(determined via `nr_php_explain_mysql_query_is_explainable()`),
otherwise it is just ignored. Later when a slow SQL query is detected
the SQL string is retrieved and used to explain the query.

An additional factor is when a `mysqli::stmt` object is released it goes
into a free pool PHP keeps which allows it to quickly hand out an
allocated section of memory when a new object is created. Each object
has a `object ID` (referenced as a `handle` in the agent for the
relevant code). When an the allocated object is reused from the free
pool it retains the same `object ID`.

The agent used the `object ID` (`handle`) as the key for storing the
prepared SQL string for explainable SQL strings. Now what if an
explainable query comes through, the string is stored, and then that
object is released. A new query is prepared with a string that is NOT
explainable and gets the object container with the same `object ID` as
the explainable one just released. The string stored in `mysqli_queries`
using that `object ID` as the key is now stale. If the new,
unexplainable query is slow then the agent will pull this stale string
and try to run an explain with it - leading to the warning.

The fix is to clear any query string for a given `object ID` if the
query string is unexplainable.

Integration tests were added (based on the reproduction case) that test
for a regression of this fix.