fixup! 29f43b9b3798e0aaefe49f8acf449b4611b1a71e
Env var substituting happens in integration_runner process, and in order
for it to work, env vars available to the test file, which executes in
PHP process, should be available to integration_runner. Therefore both
environments - integration_runner and PHP's (tests) - are queried for
value of the `k` used in `ENV[k]` when it is evaluated. In order to
preserve the behavior, integration_runner's environment is queried first
and if the value is found, it is used.