fixup! 29f43b9b3798e0aaefe49f8acf449b4611b1a71e
Env var substituting happens in integration_runner process, and in order
for it to work, env vars available to the test file, which executes in
PHP process, should be available to integration_runner. Therefore both
environments - integration_runner and PHP's (tests) - need to be made
available to integration_runner. In order for that to happen, Test.Env,
which is available to integration_runner, must include all env vars
available to PHP. Therfore PHP's env is merged into Test.Env.