fix(agent): end guzzle's external segment for rejected request (#883)
Modify the onRejected handler to end guzzle's external segment even when
there's no response. This allows to mark the span as an external span and
capture the uri and method of the request. Additionally when sync requests is
rejected, onRejected callback is not called via `PromiseInterface::then` and
needs to be called manually.

Fixes #320.

---------

Co-authored-by: Amber Sistla <asistla@newrelic.com>