fix(agent): Slim txn naming updates 1) the scheme for route run to correlate with proper OAPI naming was incorrectly listed 2) kept same naming scheme in route:run by putting it before wrapper_call and saying ok to overwrite 3) number 2 needed to incorporate fallback naming scheme 4) added fallback naming scheme (since some errors and middleware intervention hides the txn name) using dispatcher:dispatch.
When there is an exception thrown in a route that is handled by the slim routing and error middleware, it is superseded by:
Uncaught exception 'Slim\\Exception\\HttpNotFoundException' with message 'Not found.' and transformed into an http 404 error and is shown in the http.statuscode field as 404.  If the middleware isn't loaded, it is recorded as an uncaught exception and status code is 200.

Wrap function Slim\\Routing\\Dispatcher::dispatch
It will name set the txn_name in case we hit an error and Slim\\Routing\\Route::run isn't called.  It is set to be overwritable do that if we encounter Slim\\Routing\\Route::run, it can overwrite with the most accurate info (this is only relevant in the case of redirects).  This allows us to give the txn_name even in error cases when Slim\\Routing\\Route::run isn't called.