feat(agent): Add basic aws-sdk-php intrumentation (#932)
1) Detects the use of aws-sdk-php 3 library with a magic file 
2) wraps a function to extract version number
3) Sends a supportability metric for the library and package 
4) Sends PHP Package info for version
5) Multiverse tests were added in support of this

further updated to:
1. added unit tests
2. created a wrapped function that will create a metric for all detected
service names. as soon as the client is initialized, the metric is
automatically created.

Then the metric is generated as seen in the associated multiverse tests.

Please see the challenge with instrumenting for PHP 8.2+ and various
options that were considered below.
Option 2 is the currently chosen method due to lower risk, complexity,
and overhead.

Challenges:
Php8.2+ and composer autoload leads to the main magic file being
optimized directly to opcache.

Options considered:
1) for PHP8.2, and only optimizable libraries, when encountering
autoload.php files, ask the file what includes it added and check
against only the optimizable library. Small overhead incurred when
encountering an autoload file.

summary:
Magic file = `{"AWS-SDK-PHP", NR_PSTR("aws-sdk-php/src/functions.php"),
nr_aws_sdk_php_enable, true},`

Requires bool added to library structure:

```
typedef struct _nr_library_table_t {
  const char* library_name;
  const char* file_to_check;
  size_t file_to_check_len;
  nr_library_enable_fn_t enable;
  bool optimizable;
} nr_library_table_t;
```

```
Additional complexity in `nr_execute_handle_library
#if amber && ZEND_MODULE_API_NO >= ZEND_8_2_X_API_NO
/*
 * Only check if files have been optimized out for PHP 8.2+ and if we know the file is optimizable.
 * It's basically preloading things into the opcache without setting the preload flag that
 * we check to see if we need to check the opcache.
 * In this case, we need to check the PHP list of files that are included to see if our magic
 * file is there. We further limit overhead by only checking when it is an autoload file.
 */
 if ((libraries[i].optimizable) && (nr_striendswith(STR_AND_LEN(filename), "autoload.php", 12)) )  {
      zend_string *zstring_fname;
      
      if (NULL != &EG(included_files)) {
        ZEND_HASH_MAP_FOREACH_STR_KEY(&EG(included_files), zstring_fname) {
              if (nr_striendswith(ZEND_STRING_VALUE(zstring_fname), ZEND_STRING_LEN(zstring_fname) ,
                                  STR_AND_LEN(libraries[i].file_to_check))) {
                nrl_debug(NRL_INSTRUMENT, "detected library=%s",
                          libraries[i].library_name);
                
                nr_fw_support_add_library_supportability_metric(NRPRG(txn), libraries[i].library_name);
                if (NULL != libraries[i].enable) {
                  libraries[i].enable();
                }
                
  
              }  
          
    
        } ZEND_HASH_FOREACH_END();
 }
#endif // PHP 8.2+     
```
`


2) use a file that gets called later and only when AwsClient.php file
file is called. It's called later and we'll miss some instrumentation,
but if we're only ever going to be interested in `Client` calls anyway,
maybe that's ok? Doesn't detect Sdk.php (optimized out) so when
customers only use that or when they use it first, we will not
instrument it. This only detects when a Client is called to use a
service so potentially misses out on other instrumentation and misses
out when customers use the aws-sdk-php but use non-SDK way to interact
with the service (possibly with redis/memcached). This way is definitely
the least complex and lowest overhead and less complexity means lower
risk as well.

3) Just directly add the wrappers to the hash map. With potentially
50ish clients to wrap, this will add overhead to every hash map lookup.