feat(agent): optionally pad agent generated trace_ids (#929)
When the agent generates trace id based on the information it received in HTTP
headers, trace id will have 32 characters. However when the agent generates the
trace id for internal use only, the trace id has 16 characters. This PR adds an
INI setting that when set, the agent will left pad its internally used trace id
with '0' so that the trace id has 32 characters.

---------

Co-authored-by: Michal Nowacki <mnowacki@newrelic.com>