fix(testing): Fixes count check for apdex metrics (#946)
The current code for EXPECT_METRICS_EXIST would verify the count of any
metric was >0 before passing the test. However, with apdex metrics the
first field is not the count of metrics but the count of the number of
times the transaction time was less than apdex_t. This change allows for
the count to be 0 for apdex metrics.