fix(agent): Fixes newrelic_notice_error() API for PHP 8+ (#960)
Starting with PHP 8.0 the `set_error_handler()` callback function
signature drops the `$errcontext` final parameter so it only contains 4
parameters. This caused using `newrelic_notice_error()` as a callback
handler for PHP 8+ to not work as the API call did not accept only 4
arguments. This PR adds support for this function signature. There are
also numerous tests added that will hopefully caught this kind of issue
in the future.

---------

Co-authored-by: Amber Sistla <asistla@newrelic.com>
Co-authored-by: Hitesh Ahuja <108540135+hahuja2@users.noreply.github.com>