fix(agent): Changes package major metric creation
Changes the creation of package major metrics to occur at the end
of a transaction.  Packages are marked to have a metric created using
the new function nr_txn_php_package_set_options().  If the option
NR_PHP_PACKAGE_OPTION_MAJOR_METRIC is set on a package  then
at the end of the transaction a major supportability metric will be
be created for that package.