feat(agent): Moves major version metric creation to R_SHUTDOWN
Changes how the package major number metrics are created so this occurs
in R_SHUTDOWN.  This allows the use of package versions from ALL sources
include the Composer API.  Whenever instrumentation for a package detects
a package it can create a package suggestion.  Initially if a version is not
known it is fine to use PHP_PACKAGE_VERSION_UNKNOWN.  If latest
the instrumentation determines a version from a class constant, etc, then
the suggestion can be updated with the version.

At the end of the transaction then this list of suggestions is traversed and
the actual package data (which could include Composer API data) is
referenced and a major number supportability metric is created with the
best version available.