fix(agent): Don't skip arguments when calling mysqli::real_connect
When one of the arguments in nr_php_mysqli_link_real_connect is
null, the entire argument is skipped. This causes a fatal error
when attempting to connect to a database through a Unix socket,
without specifying a port number.

This change ensures all arguments are passed on to real_connect
in the correct order, including any nulls.