fix(installer): tarball installer checks for ini in mods-available (#971)
Debian uses a directory called "mods-available" to house ini files. It's
`phpenmod` and `phpdismod` allow command-line management of extensions
in directory. It also creates a symlink ini file to this
"mods-available" directory in the normal conf.d directory. The issue is
that this symlink is prefixed with a number (for example
20-newrelic.ini). This prefix prevents our tarball installer from
recognizing that a newrelic.ini already exists, resulting in 2 ini
files: a newrelic.ini and a 20-newrelic.ini.

Our .deb package installer scouts out the "mods-available" directory and
installs therein when possible. That means that if a customer installs
via package but upgrades via tarball (not recommended), then their
system gets the duplicates.

This change makes our tarball installer scout for a potential previous
installation in mods-available.

Fixes https://github.com/newrelic/newrelic-php-agent/issues/399

---------

Co-authored-by: Michal Nowacki <mnowacki@newrelic.com>