fix(agent): only non-null for first three params
Part of the problem is that argc in the function previously only incremented for non-null parameters, so it's throwing the count off when we are using it in if statements.  Additionally, for our instrumentation to work, the first three parameters should only be added if they are non-null (note the comment that exists before adding the remaining params). So we need to keep track of the required first three in addition to all other arguments that are being added.  We also need the count of the required parameters so that we can properly assess whether we need to do call call mysqli::select_db. We can handle this with an additional param to specify whether a value should be added as an arg if it is set or not set.