eliminate filename case conversion
Only filename's suffix needs to be lowercase because only the suffix is used
for key's hash. This observation leads to optimizing out nr_strdup done by
nr_string_to_lower. Also only small portion of the filename needs its case
to be converted, not the whole filename (with path, which can be long).