don't fallback onto deps from vendor
Abort build if protobuf-c cannot be found in the system without falling
back to building and using protobuf-c from vendor subdir. axiom build
needs protoc-c (provided by protobuf-c-compiler package) and agent needs
to link to protobuf-c static library (provided by protobuf-c-devel).
Static library is required to avoid shared objects runtime dependencies
on customers' systems.

Don't attempt to build protobuf compiler from vendor subdir - assume it
is installed and let the build fail if it is not.  protobuf compiler
alone is not enough for the rule that used it anyway - protoc-gen-go is
needed too, and it has never been vendored but rather was assumed to be
installed on the build system. The rule that uses protobuf compiler and
protoc-gen-go is not part of the build anyway - its output is checked in
to the repository. The rule exists as documentation how to regenerate
infinite tracing code for the daemon.