fix: add fallback injection logic
This change handles the edge case exposed in our tests where the agent
may fail to inject the cache naming function at enable because it looked
for drupal and symfony classes before composer autoload had time to load
them. In the normal flow, this would only be a temporary problem since
nr_drupal_enable should be re-run every request, and subsequent requests
should have these resources available. Our test setup, however, shows
that if in the process of handling one request, another request is
generated and the agent doesn't re-trigger RINIT processing, the
function name from the first request will bleed through to the second
request and result in an incorrect name.
To fix, extract logic responsible for injecting the cache naming
function to its own function and call both at enable and within the
PageCache::get function instrumentation, provided the injection function
(newrelic_name_cached) is not detected when we look for it. This should
strike the best balance between performance and correctness.