refactor wraprec storage and lookup [wip]
Implement wraprec lookup by scope and function name. The implementation uses
two specialized hashmaps: one for global functions and another one for scopes.
The scopes hashmap's values are hashmaps for scoped methods. The hashmaps use
keys composed of not only the string and its length but also zend string's
hash. This hash is used to find the bucket index. The decision to include zend
string's hash was deliberate because the agent gets it for free when it needs
to lookup the wraprec based on scope and function name that are available in
execute_data's op array as zend_strings.