fix(daemon): Properly escape UTF-8 in log labels
The previous code did not account for that labels can contain UTF-8 characters which
must be escaped.  This commit changes the code to use the Go marshalling for json
which will properly handle these characters.  A guard was added to make sure invalid
log labels are not converted to JSON (with an empty key or value).  The ingest code
should prevent this from happening but a secondary check here on JSON generation
is a reasonable safeguard.