fix(daemon): Adds escaping for log forwarded labels (#1035)
This PR corrects how the collector JSON for log forwarded labels handles
log label key and values which have characters which must be escaped for
JSON marshaling. The change is restricted to the daemon code to generate
the log label JSON and does not impact the agent side of things.

The fix is to use the Go `json` module to handle the creation of the
JSON. It will handle the escaping as needed. The array of label
key/value pairs is converted to a Go `map` which is then marshaled into
JSON. The case where either the label or value is empty (an invalid
label) is handled so that these are not forwarded. This case should not
occur as labels of this type should not be created by the agent but this
check is added as additional protection against this case.