seventh round of fixes
Adjust the invokeAllWith test to a different functionality of the agent - first
installed instrumentation wins. Adding a named wraprec != installing special
instrumentation. With observer API special instrumentation (wraprec) is
installed in zend function's op_array extension's handle at the fcall_init,
when the function is executed for the first time, not at the time when wraprec
is created, which is just to associate name of the function with wraprec to
look it up at fcall_init.