fixup! 986ba9dd36f97694c7c7c11c91b261e9e6221047
Agent way of installing wraprecs when running in PHPs not using observer API
(< 8.0) requires file to be executed **after** the function is loaded by PHP.
When custom tracer is added via configuration, the function for which the custom
tracer (wraprec) is added, is not loaded by PHP yet. Only file execution
triggers re-scanning list of custom tracers and installing them.