refactor(agent): optimize user function instrumentation lookup for PHPs 8.0+ (#1042)
Reduce agent’s CPU overhead when observer API is used to hook
into Zend Engine. This is achieved by:
* Simplifying the checks the agent needs to perform to determine if
special instrumentation for user function is provided. Before this
change, on every function execution the agent used a hashmap to check if
special instrumentation for user function is provided. However, when the
observer API is used to hook into Zend Engine, this can be done only
once per user function execution at the time when the user function is
executed for the first time.
* Reducing the number of times the name of the file is compared with the
list of ‘magic files’ to detect package and/or framework used. Before
this change on every file execution the agent scanned the list of magic
files twice. However, when the observer API is used to hook into Zend
Engine, this can be done only once per file execution.

---------

Co-authored-by: Amber Sistla <asistla@newrelic.com>